refactor: Reorganize ContentItem into layered packages
Phase 2 of the restructure: split the flat contentitem package by
concern, and lift content-item-agnostic infrastructure up to top-level
packages so other content types can share it later.

Top-level (generic):
- common: UuidV7
- persistence: JdbcConfig, SchemalessData
- elasticsearch: ElasticSearchConfig, ElasticSearchIndexCreator,
  IndexController, SearchResponse(De)serializer

contentitem (domain):
- root: ContentItemService, ContentItemRepository
- rest: ContentItemController, ContentItemExceptionHandler
- exceptions: ContentItemNotFoundException, InvalidContentItemException
- model: ContentItemEntity, the DTOs, ContentItemMapper
- transformation: the inbound transformers + TransformationHandler
- elasticsearch: the content-aware indexer, search controller, and
  ESRecordTransformers

Pure move + import bookkeeping; no behaviour change. CLAUDE.md documents
the new layout.
