+++ src/main/java/com/contented/contented/contentitem/ContentItemRepository.java
@@ -1,5 +1,6 @@
 package com.contented.contented.contentitem;
 
+import com.contented.contented.contentitem.model.ContentItemEntity;
 import org.springframework.data.repository.ListCrudRepository;
 
 import java.util.UUID;
