+++ src/main/java/com/contented/contented/contentitem/elasticsearch/ContentItemIndexer.java
@@ -1,6 +1,6 @@
 package com.contented.contented.contentitem.elasticsearch;
 
-import com.contented.contented.contentitem.ContentItemEntity;
+import com.contented.contented.contentitem.model.ContentItemEntity;
 import com.contented.contented.contentitem.elasticsearch.transformation.ESRecordTransformer;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
