+++ src/main/java/com/contented/contented/contentitem/elasticsearch/SearchController.java
@@ -3,7 +3,7 @@ package com.contented.contented.contentitem.elasticsearch;
 import co.elastic.clients.elasticsearch.ElasticsearchClient;
 import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.SearchResponse;
-import com.contented.contented.contentitem.ContentItemResponseDTO;
+import com.contented.contented.contentitem.model.ContentItemResponseDTO;
 import com.contented.contented.contentitem.ContentItemService;
 import org.springframework.data.elasticsearch.client.elc.EntityAsMap;
 import org.springframework.data.elasticsearch.core.mapping.IndexCoordinates;
