+++ src/main/java/com/contented/contented/contentitem/elasticsearch/SearchResultsWithContent.java
@@ -1,7 +1,7 @@
 package com.contented.contented.contentitem.elasticsearch;
 
 import co.elastic.clients.elasticsearch.core.SearchResponse;
-import com.contented.contented.contentitem.ContentItemResponseDTO;
+import com.contented.contented.contentitem.model.ContentItemResponseDTO;
 
 import java.util.List;
