+++ src/main/java/com/contented/contented/contentitem/elasticsearch/transformation/StandardContentItemTransformations.java
@@ -1,6 +1,6 @@
 package com.contented.contented.contentitem.elasticsearch.transformation;
 
-import com.contented.contented.contentitem.ContentItemEntity;
+import com.contented.contented.contentitem.model.ContentItemEntity;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.elasticsearch.client.elc.EntityAsMap;
