+++ src/main/java/com/contented/contented/contentitem/model/ContentItemMapper.java
@@ -1,4 +1,4 @@
-package com.contented.contented.contentitem;
+package com.contented.contented.contentitem.model;
 
 import org.springframework.stereotype.Component;
