+++ src/main/java/com/contented/contented/contentitem/model/ContentItemResponseDTO.java
@@ -1,4 +1,4 @@
-package com.contented.contented.contentitem;
+package com.contented.contented.contentitem.model;
 
 public class ContentItemResponseDTO extends AbstractContentItemDTO {
 }
