+++ src/main/java/com/contented/contented/contentitem/transformation/Transformer.java
@@ -1,6 +1,6 @@
 package com.contented.contented.contentitem.transformation;
 
-import com.contented.contented.contentitem.ContentItemEntity;
+import com.contented.contented.contentitem.model.ContentItemEntity;
 
 public interface Transformer<T, R> {
