+++ src/test/java/com/contented/contented/contentitem/ContentItemControllerFieldTests.java
@@ -1,6 +1,8 @@
 package com.contented.contented.contentitem;
 
 import com.contented.contented.contentitem.elasticsearch.ContentItemIndexer;
+import com.contented.contented.contentitem.model.ContentItemEntity;
+import com.contented.contented.contentitem.model.ContentItemResponseDTO;
 import com.contented.contented.contentitem.testutils.StubbingUtils;
 import org.junit.jupiter.api.*;
 import org.springframework.boot.test.context.SpringBootTest;
