+++ src/test/java/com/contented/contented/contentitem/ContentItemMapperTest.java
@@ -1,5 +1,10 @@
 package com.contented.contented.contentitem;
 
+import com.contented.contented.common.UuidV7;
+import com.contented.contented.contentitem.model.ContentItemDTO;
+import com.contented.contented.contentitem.model.ContentItemEntity;
+import com.contented.contented.contentitem.model.ContentItemMapper;
+import com.contented.contented.contentitem.model.ContentItemResponseDTO;
 import com.contented.contented.contentitem.testutils.NestedPerClass;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
