+++ src/test/java/com/contented/contented/contentitem/ContentItemServiceTest.java
@@ -1,9 +1,16 @@
 package com.contented.contented.contentitem;
 
+import com.contented.contented.common.UuidV7;
 import com.contented.contented.contentitem.elasticsearch.ContentItemIndexer;
 import com.contented.contented.contentitem.elasticsearch.transformation.BlogTransformer;
+import com.contented.contented.contentitem.exceptions.InvalidContentItemException;
 import com.contented.contented.contentitem.testutils.NestedPerClass;
+import com.contented.contented.contentitem.model.ContentItemDTO;
+import com.contented.contented.contentitem.model.ContentItemEntity;
+import com.contented.contented.contentitem.model.ContentItemMapper;
+import com.contented.contented.contentitem.model.ContentItemResponseDTO;
 import com.contented.contented.contentitem.transformation.StandardDMSContentTransformer;
+import com.contented.contented.contentitem.transformation.TransformationHandler;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
