+++ src/test/java/com/contented/contented/contentitem/elasticsearch/SearchControllerTests.java
@@ -2,7 +2,9 @@ package com.contented.contented.contentitem.elasticsearch;
 
 import co.elastic.clients.elasticsearch.core.SearchResponse;
 import com.contented.contented.contentitem.AbstractContentItemControllerTests;
-import com.contented.contented.contentitem.ContentItemResponseDTO;
+import com.contented.contented.contentitem.model.ContentItemResponseDTO;
+import com.contented.contented.elasticsearch.ElasticSearchIndexCreator;
+import com.contented.contented.elasticsearch.SearchResponseDeserializer;
 import com.contented.contented.contentitem.testutils.NestedPerClass;
 import tools.jackson.databind.annotation.JsonDeserialize;
 import org.junit.jupiter.api.*;
