+++ src/test/java/com/contented/contented/contentitem/elasticsearch/SearchResponseDeserializerTests.java
@@ -1,6 +1,7 @@
 package com.contented.contented.contentitem.elasticsearch;
 
 import co.elastic.clients.elasticsearch.core.SearchResponse;
+import com.contented.contented.elasticsearch.SearchResponseDeserializer;
 import tools.jackson.databind.DatabindException;
 import tools.jackson.databind.annotation.JsonDeserialize;
 import tools.jackson.databind.json.JsonMapper;
