+++ src/test/java/com/contented/contented/contentitem/transformation/StandardDMSContentTransformerTest.java
@@ -1,7 +1,7 @@
 package com.contented.contented.contentitem.transformation;
 
-import com.contented.contented.contentitem.ContentItemEntity;
-import com.contented.contented.contentitem.UuidV7;
+import com.contented.contented.contentitem.model.ContentItemEntity;
+import com.contented.contented.common.UuidV7;
 import com.contented.contented.contentitem.testutils.NestedPerClass;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
